rename app constants and files to match rails 7 zentiwerk constant, replace .keys with attribute_names
